requirements.txt
streamlit==1.32.0
tensorflow==2.17.0
Pillow==10.1.0
numpy==1.26.0
